pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tldr"
version = "0.1.0"
dependencies = [
    "yt-dlp>=2025.06.25",
    "ffmpeg-python>=0.2.0",
    "faster-whisper>=0.10.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "numpy<2", # pin to numpy 1.x
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tldr*"]
exclude = ["downloads*", "audio*"]
